re-add profile fetch
