+++ src/actions/picActions.js
@@ -75,31 +75,3 @@ export const editLikes = (pic) => {
     );
   }
 }
-
-export const addFriend = (friend) => dispatch => {
-  const token = localStorage.token;
-  fetch (`${url}/friendships`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      Accept: 'application/json',
-      'Authorization': `Bearer ${token}`
-    },
-    body: JSON.stringify (friend)
-  })
-    .then (resp => resp.json())
-    .then (data => {
-      if (data.error) {
-        dispatch({
-          type: 'ADD_FRIEND_FAIL',
-          payload: data.error
-        })
-      } else {
-        dispatch({
-          type: "ADD_FRIEND",
-          payload: data
-        })
-      }
-    }
-  );
-}
