+++ src/components/SingleStation.js
@@ -1,76 +1,94 @@
-import React, { Component } from 'react'
-import { connect } from 'react-redux';
-import defaultImg from '../images/404.jpg';
-import { fetchStations, editLikes } from '../actions/stationActions';
-import Comments from './Comments'
-import CommentForm from '../components/CommentForm'
-import { MdFavorite } from "react-icons/md";
-import { MdFavoriteBorder } from "react-icons/md"
+import React, { useEffect, useState } from 'react';
+import { useDispatch, useSelector } from "react-redux";
+import Dialog from "@material-ui/core/Dialog";
+import IconButton from '@material-ui/core/IconButton';
+import Toolbar from '@material-ui/core/Toolbar';
+import Typography from '@material-ui/core/Typography';
+import CloseIcon from '@material-ui/icons/Close';
+import Slide from '@material-ui/core/Slide';
+import { MdFavorite, MdFavoriteBorder } from "react-icons/md";
+import StationMap from "./StationMap";
+import Comments from "./Comments";
+import CommentForm from "./CommentForm";
 
-import StationMap from '../components/StationMap'
+import { fetchSingleStation, editLikes } from "../actions/stationActions";
+import { AppBar } from "@material-ui/core";
 
-class SingleStation extends Component {
+const Transition = React.forwardRef(function Transition(props, ref) {
+  return <Slide direction="up" ref={ref} {...props} />;
+});
 
-    componentDidMount(){
-        this.props.fetchStations()
-    }
+const SingleStation = ({ open, setOpen, stationId }) => {
+  const [liked, setLiked] = useState(false);
 
-    state = {
-        liked: false,
-        pending: true
-    }
+  const dispatch = useDispatch();
+  const comment = useSelector(state => state.comment);
+  const station = useSelector(state => state.stations.station);
+  const currentUser = useSelector(state => state.currentUser.currentUser);
 
-    addLikes = () => {
-        this.props.editLikes(this.props.station)
-        this.setState({liked: !this.state.liked, pending: false})
-    }
+  useEffect(() => {
+    dispatch(fetchSingleStation(stationId));
+  }, [stationId, dispatch, comment]);
 
-    render() {
-        return (
-            <div className="single-station">
-            <h1>{this.props.station.name}</h1>
-            <img className="big-pic" src={{...this.props.station.pics[0]}.pic_url ? {...this.props.station.pics[0]}.pic_url : defaultImg } alt=""/>
-            <div className="likes">
-                <div className="heart" onClick={this.state.pending ? this.addLikes : null}>
-                {this.state.liked ? <MdFavorite/> : <MdFavoriteBorder/>}
-                <h3>Likes: {this.props.station.likes}</h3>
-                </div>
-            <h3>{this.props.station.address}</h3>
-            </div>
-            <h3 className="ui dividing header">Location</h3>
-            <div className="location">
-                {this.props.station.coords === "" ?
-                <h1>Loading....</h1>
-                : <StationMap coords={this.props.station.coords}/>}
-            </div>
-            <div className="ui comments">
-            <h3 className="ui dividing header">Comments</h3>
-            <Comments comments={this.props.comments}/>
-            {localStorage.token && (
-                <CommentForm type="Station" id={this.props.station.id} />
-            )}
-            </div>
-            </div>
-        )
-    }
+  if (!station || station.status === 404 || !stationId) {
+    return null;
+  }
+
+  const addLikes = () => {
+    setLiked(true);
+    dispatch(editLikes(station));
+  }
+
+  const handleClose = () => {
+    setOpen(false);
+  };
+
+  return (
+    <Dialog
+      fullScreen
+      open={open}
+      onClose={handleClose}
+      TransitionComponent={Transition}
+    >
+      <AppBar position="sticky" style={{ backgroundImage: 'linear-gradient(135deg, #0092D0, #FDBE43, #FF7500)' }}>
+        <Toolbar>
+          <IconButton edge="start" color="inherit" onClick={handleClose} aria-label="close">
+            <CloseIcon />
+          </IconButton>
+          <Typography variant="h2" style={{ fontSize: '3rem', marginLeft: '16px', flex: 1 }}>
+            {station.name}
+          </Typography>
+        </Toolbar>
+      </AppBar>
+      <div className="single-station">
+        <img className="big-pic" src={station.first_pic_url} alt={station.name} />
+        <div className="likes">
+          <div className="heart" onClick={liked ? null : addLikes}>
+            {liked ? <MdFavorite/> : <MdFavoriteBorder/>}
+            <h3>Likes: {station.likes}</h3>
+          </div>
+          <h3>{station.address}</h3>
+        </div>
+        <h3 className="ui dividing header">Location</h3>
+        <div className="location">
+          {station.coords === "" ?
+            <h1>Loading....</h1>
+            : <StationMap coords={station.coords}/>}
+        </div>
+        <div className="ui comments">
+          {station.comments && (
+            <React.Fragment>
+              <h3 className="ui dividing header">Comments</h3>
+              <Comments comments={station.comments} loading={comment.loading} />
+            </React.Fragment>
+          )}
+          {currentUser && currentUser.username && (
+            <CommentForm type="Station" id={station.id} error={comment.error} />
+          )}
+        </div>
+      </div>
+    </Dialog>
+  );
 }
-function mapStateToProps(state, ownProps){
-    let station = {id:'', name:'', slug:'', likes:'', address: '', coords: '', description: '', pics:[], users: [], comments:[]};
-    const slug = ownProps.match.params.slug;
-    if(state.stations.stations.length > 0){
-        station = Object.assign({}, state.stations.stations.find(station => station.slug === slug))
-    }
-    if(station.comments.length >= 0){
-        station.comments = station.comments.concat(state.comments.comments)
-    }
-    return {station,
-    comments: station.comments}
-}
 
-const mapDispatchToProps = (dispatch) => ({
-    fetchStations: () => dispatch(fetchStations()),
-    editLikes: (station) => dispatch(editLikes(station))
-})
-
-
-export default connect(mapStateToProps, mapDispatchToProps)(SingleStation)
+export default SingleStation;
